fix: Secured posts API endpoints
- Secured API endpoints in posts namespace
- To do the update and delete operations
- The session user and post owner should be same
- Update and delete operations are denied if it is done by other user

+++ app/API/posts/delete.php
@@ -7,14 +7,20 @@ use App\Model\Post;
 ${basename(__FILE__, '.php')} = function () {
     if ($this->isAuthenticated() && $this->get_request_method() == 'POST') {
         if ($this->paramsExists(['id']) && !empty($this->_request['id'])) {
-            $p = new Post($this->_request['id']);
-            if ($p->hasMultipleImages($this->_request['id'])) {
-                $this->response($this->json([
-                    'message'=>$p->deleteMultiplePostImages()
-                ]), 200);
+            if (Post::verifyOwner($this->_request['id'])) {
+                $p = new Post($this->_request['id']);
+                if ($p->hasMultipleImages($this->_request['id'])) {
+                    $this->response($this->json([
+                        'message' => $p->deleteMultiplePostImages()
+                    ]), 200);
+                } else {
+                    $this->response($this->json([
+                        'message' => $p->deleteSinglePostImage()
+                    ]), 200);
+                }
             } else {
                 $this->response($this->json([
-                    'message'=>$p->deleteSinglePostImage()
+                    'message' => false
                 ]), 200);
             }
         } else {
